apps/autointerp/pyproject.toml
[tool.poetry]
name = "neuronpedia_autointerp"
version = "1.0.0"
description = "Neuronpedia's AutoInterp Server - Based on EleutherAI's sae-auto-interp."
authors = ["Johnny Lin <johnny@neuronpedia.org>"]
license = "MIT"
readme = "README.md"
keywords = ["interpretability", "explainable-ai"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.115.0"
pydantic = "^2.0.0"
uvicorn = "^0.34.0"
python-dotenv = "^1.1.0"
neuronpedia-autointerp-client = {path = "../../packages/python/neuronpedia-autointerp-client"}
sentry-sdk = { extras = ["fastapi"], version = "^2.20.0" }
eai-delphi = "0.0.2"
bitsandbytes = "==0.42.0"
torchtyping = "^0.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
coverage = "^7.6.0"
ruff = "^0.7.4"
pyright = "^1.1.351"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"typing.Union".msg = "Use '|' instead"
"typing.Optional".msg = "Use '| None' instead"
"typing.Dict".msg = "Use 'dict' instead"
"typing.List".msg = "Use 'list' instead"

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = "none"
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
reportUnknownVariableType = "none"
reportUntypedFunctionDecorator = "none"
reportUnnecessaryIsInstance = "none"
reportUnnecessaryComparison = "none"
reportConstantRedefinition = "none"
reportUnknownLambdaType = "none"
reportUnknownParameterType = "none"
reportPrivateUsage = "none"
